Printer et Logos pour les plannings
- Ajout du printer
- Ajout support des logos sur les plannings
- Correction de bugs

+++ v3.0/planning/del_planning.php
@@ -9,19 +9,41 @@
   	{
   		$typePL = $_POST['typePL'];
   		$date=DateTime::createFromFormat('d/m/Y', $_POST["Date"])->format('Y-m-d');
-		$query = mysqli_query($db, 'DELETE FROM pl_association WHERE ASSOC_date="'.$date.'" AND PL_id = '.$typePL.';');
-  		if($query)
-		{
-	        echo '<div id="good">     
-	            <label>Le planning de la semaine du lundi '.date("d/m/Y", strtotime($date)).' supprimé avec succès !</label>
-	            </div>';
+        if(mysqli_query($db, 'SET autocommit=0;') && mysqli_query($db, 'START TRANSACTION;'))
+        {
+            $query = mysqli_query($db, 'DELETE FROM pl_association WHERE ASSOC_date="'.$date.'" AND PL_id = '.$typePL.';');
+            if($query)
+            {
+                $query = mysqli_query($db, 'DELETE FROM logo_association WHERE ASSOC_date="'.$date.'" AND PL_id = '.$typePL.';');
+                if($query)
+                {
+                    mysqli_query($db, 'COMMIT;');
+                    echo '<div id="good">     
+                        <label>Le planning de la semaine du lundi '.date("d/m/Y", strtotime($date)).' supprimé avec succès !</label>
+                        </div>';
+                }
+                else
+                {
+                    mysqli_query($db, 'ROLLBACK;');
+                    echo '<div id="bad">     
+                          <label>Une erreur s\'est produite lors de la suppression du planning !</label>
+                          </div>';
+                }
+            }
+            else
+            {
+                mysqli_query($db, 'ROLLBACK;');
+                echo '<div id="bad">     
+                      <label>Une erreur s\'est produite lors de la suppression du planning !</label>
+                      </div>';
+            }
         }
-		else
-		{
-			echo '<div id="bad">     
-			      <label>Une erreur s\'est produite lors de la suppression du planning !</label>
-			      </div>';
-		}
+        else
+        {
+            echo '<div id="bad">     
+                  <label>Une erreur s\'est produite lors de la suppression du planning !</label>
+                  </div>';
+        }
 	}
 	else
 	{
